fix some bugs with table of gestures

+++ tools/MouseGestures/pathcorrector.h
@@ -9,7 +9,7 @@ public:
     static bool isLine(QList<QPoint> const & path);
     static QList<QPoint> getMousePath(QList<QPoint> const & path);
     static QList<QPoint> rotate(QList<QPoint> const & path);
-    static QList<QPoint> increase(QList<QPoint> const & path, int k);
+    static QList<QPoint> increase(QList<QPoint> const & path, double k);
 
 private:
     static double getCos(QPoint const & pnt1, QPoint const & pnt2, QPoint const & pnt3);
